# src/OpenApi/OpenApi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OpenApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://dev.mcdonaldconsulting.net"
    }
  ],
  "paths": {
    "/Customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Get all customers",
        "description": "Retrieves all customers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.6.1",
                  "title": "An error occurred while processing your request.",
                  "status": 500,
                  "traceId": "01234567-1234-0123-0123-0123456789AB"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a new customer",
        "description": "Creates a new customer in the system.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.6.1",
                  "title": "An error occurred while processing your request.",
                  "status": 500,
                  "traceId": "01234567-1234-0123-0123-0123456789AB"
                }
              }
            }
          }
        }
      }
    },
    "/Customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Get a customer by ID",
        "description": "Retrieves a customer by their unique identifier.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                },
                "example": {
                  "id": 1,
                  "name": "John Doe"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.6.1",
                  "title": "An error occurred while processing your request.",
                  "status": 500,
                  "traceId": "01234567-1234-0123-0123-0123456789AB"
                }
              }
            }
          }
        }
      }
    },
    "/Customers/error": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Demonstrate error handling",
        "description": "This endpoint is used to demonstrate error handling in the API.",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.6.1",
                  "title": "An error occurred while processing your request.",
                  "status": 500,
                  "traceId": "01234567-1234-0123-0123-0123456789AB"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier for the customer.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "The name of the customer."
          }
        },
        "example": {
          "id": 1,
          "name": "John Doe"
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "example": {
          "type": "https://tools.ietf.org/html/rfc9110#section-15.6.1",
          "title": "An error occurred while processing your request.",
          "status": 500,
          "traceId": "01234567-1234-0123-0123-0123456789AB"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Customers"
    }
  ]
}